Showing Succes Message for Updating Users
Defect: After updating user details, page is redirected to users main page but there was no success message to indicate whether user was updated or not.

This change addresses the above defect.
In manageUserController.js, I have added a flag to indicate update and used it in usersController.js for showing Success Message.

In view.html I have added the div element that contains the success message.

HPQC Defect: 22430