Adding missed Search LOV parameters in Contact Requests grid.
Defect - 24483

The search LOV in this tab was missing most of the parameters that are present  in the final grid columns document.
The missed parameters have been added in the Search LOV.

Changes made:
serviceRequestsFactory.js - Have removed the 'notSearchable' attribute for all the columns and added the missed column field names in the grid.